ARM: msm: Rename PMIC pm8644 to pma8084
Rename PMIC pm8644 to pma8084 in filenames, node names, property
values, and phandles.  PMIC PM8644 has officially been renamed to
PMA8084 so this should be reflected in device tree.

At the same time, fix up the PMA8084 GPIO, MPP, and regulator
nodes so that they match what is actually present in the PMIC.

Change-Id: If6bf47e1c5501d3ece3d9dc45a482470e4665597
Signed-off-by: David Collins <collinsd@codeaurora.org>