usb: usb_bam: add bam global inactivity and activity notifications
Bam global inactivity and activity notifications are notifications
that usb bam driver can notify its clients. Global activity notification
occur when data is pulled or pushed on at least one of the bam pipes
which its direction is into the usb bam. Global inactivity notification
occur when no data is pulled or pushed on any of the bam pipes,
in both directions. The usb bam driver toggle between the activity
and inactivity states, so successive notifications will not be from
the same type.

This change uses dedicated interrupts from the bam hardware.
For example the global inactivity timer is hardware feature
that allows to get an interrupt upon idle of all the bam pipes
for some predefined duration.

This feature is required in some scenarios in which we want to detect
activity and no activity on the bam pipes and take action,
for example exit or enter into power save modes.

Change-Id: Ib0f51ca59456a642dcdb921784fdcbcac3d552ca
Signed-off-by: Ido Shayevitz <idos@codeaurora.org>