board: 8610: add SD card detect gpiomux configuration
On MSM8x10 platforms, a MSM GPIO is used for SD card detect and it's
gpiomux configuration needs to be set to a specified value for the
SD card to be detected across all platforms. This patch adds the correct
gpiomux configuration for SD card detect GPIO to properly detect SD card
across all platforms.

CRs-Fixed: 480392
Change-Id: I4ce31b329ec3ff1d9dcad7355851cb28beffcd82
Signed-off-by: Venkat Gopalakrishnan <venkatg@codeaurora.org>