crypto: msm: Fix driver crash when running AES-CBC decryption
The cause of the crash is due to uninitialized variable "areq" used in
ablk_cipher_req().

When the areq is pulled from pce_dev->areq, it is not valid and is
uninitialized. Since we already have the valid pointer to areq from
qce_req structure, we do not need to redefine areq.

Change-Id: Ib63ec65dc9df5f87ff470c04cfe5fbb66c19199a
Signed-off-by: Mona Hossain <mhossain@codeaurora.org>