msm: subsystem_restart: Change the is_loadable flag to is_not_loadable
The is_loadable flag introduced in the subsystem descriptor is not used
by all subsystems. For subsystems not using this flag, the behavior
must be same as before introducing this flag. Do this by changing
the flag to is_not_loadable. So only subsystems without PIL will set
this flag. Subsystems using PIL will not set this flag.

Change-Id: I44c056660a2cc4df9c86b81db9c3983092519107
Signed-off-by: Seemanta Dutta <seemanta@codeaurora.org>