power: qpnp-charger: add OTG and boost regulator support
Add support for OTG and SMBB_BOOST regulators. This allows
for the relationship between the OTG FET and BOOST to be
correctly managed depending on board specifics using
the regulator framework.

The boost regulator voltage can be configured between
4.2 and 5.5V. It is recommended to set it to 5 V
which is the standard output voltage for USB OTG.

The OTG regulator simply sets the OTG_EN bit in the
USB_CHGPTH peripheral to ensure that the USB path
is correctly configure to not draw current and enable
the OTG OVP FET.

CRs-Fixed: 486088
Change-Id: I365b54cc471adac06bdcd19d02c19107f7aedd94
Signed-off-by: David Keitel <dkeitel@codeaurora.org>