gpu: ion: Copy client name into trace event
ftrace saves the information from events into a buffer to
be parsed out later. If the Ion client dies between the time
the event happens and the events are parsed, the reference to
the client name may have become invalid. Fix this by making a
local copy of the client name for event parsing.

Change-Id: I608103c763526aef0e8dd5a563be1949643e4199
Signed-off-by: Laura Abbott <lauraa@codeaurora.org>